+++ Changelog.md
@@ -81,6 +81,7 @@
 `SizeRelativeVertical`, `OOXMLShapeBodyProperties`, `Tile`,
 `NaryLimitLocation`, `MultilevelType`
 * Replace usage `Alignment.parse` on `OOXMLDocumentObject#value_to_symbol` method
+* Refactor `Color.from_int16` to `Color#parse_hex_string` method
 
 ## 0.1.2 (2016-06-07)